pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "d-pygen"
version = "0.1.4"
description = "AI-powered project generator CLI with plugin marketplace"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.9"
license = { file = "LICENSE" }

authors = [
  { name = "Sayan Bose", email = "sayanbose2004@gmail.com" }
]

keywords = [
  "ai",
  "cli",
  "code-generation",
  "project-generator",
  "templates",
  "plugins"
]

classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Environment :: Console",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Code Generators",
  "Topic :: Utilities"
]

dependencies = [
  "rich>=13.0",
  "requests>=2.28",
  "python-dotenv>=1.0",
  "openai>=1.0",
  "pyfiglet>=1.0"
]


[project.optional-dependencies]

openrouter = [
  "openai>=1.0"
]

ollama = [
  "requests>=2.28"
]

all = [
  "openai>=1.0",
  "requests>=2.28"
]


[project.urls]
Homepage = "https://github.com/dangerSayan"
Repository = "https://github.com/dangerSayan/d_pygen"
Issues = "https://github.com/dangerSayan/d_pygen/issues"


[project.scripts]
d_Pygen = "d_pygen.cli:main"


[tool.setuptools.packages.find]
where = ["."]
include = ["d_pygen*"]


[tool.setuptools]
include-package-data = true